Revert feat: support new prisma client
