--- packages/prisma-fabbrica/bin-next.js
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-require("./lib/generatorNext.js");
